auth/requirements.txt
Flask==1.1.1
mysql-connector-python==8.0.19
flask_cors
cryptography==2.9.2
pycryptodome
